feature: pass tests for sorting trips
