+++ src/scripts.js
@@ -65,6 +65,7 @@ function sortPastDests(pastTrips, dests) {
   let pastDests = allDests.filter((dest) => {
     return tripIDs.includes(dest.id)
   })
+  console.log('past', pastTrips)
   yearlyCost(pastTrips, pastDests)
   renderPastTrips(pastTrips, pastDests)
   return pastDests
@@ -78,6 +79,7 @@ function sortCurrentDests(currentTrips, dests) {
   let currentDests = allDests.filter((dest) => {
     return tripIDs.includes(dest.id)
   })
+  console.log('current', currentTrips)
   renderCurrentTrips(currentTrips, currentDests)
   return currentDests
 }
