+++ test/sample-trips.js
