+++ test/script-test.js
@@ -7,58 +7,32 @@ function userTrips(userSample, tripsSample) {
   let sortedTrips = trips.filter((trip) => {
     return userSample.id === trip.userID
   })
-  sortTrips(sortedTrips)
   return sortedTrips
 }
   
-function sortTrips(trips) {
+// function sortTrips(trips) {
+//   let pastTrips = trips.filter((trip) => {
+//     return trip.status === 'approved'
+//   })
+//   let currentTrips = trips.filter((trip) => {
+//     return trip.status === 'pending'
+//   })
+// }
+
+function sortPastTrips(trips) {
   let pastTrips = trips.filter((trip) => {
     return trip.status === 'approved'
   })
-  let currentTrips = trips.filter((trip) => {
-    return trip.status === 'pending'
-  })
-  console.log(pastTrips)
-  console.log(currentTrips)
-}
-
-let pastTrips;
-let currentTrips;
-
-function sortPastTrips(trips) {
-  return pastTrips = trips.filter((trip) => {
-    return trip.status === 'approved'
-  })
+  return pastTrips
 }
 
 function sortCurrentTrips(trips) {
-  return currentTrips = trips.filter((trip) => {
+  let currentTrips = trips.filter((trip) => {
     return trip.status === 'pending'
   })
+  return currentTrips
 }
 
-function sortPastDests(pastTrips, dests) {
-  let allDests = dests.destinations
-  let tripIDs = pastTrips.map((trip) => {
-    return trip.destinationID
-  })
-  let pastDests = allDests.filter((dest) => {
-    return tripIDs.includes(dest.id)
-  })
-  yearlyCost(pastTrips, pastDests)
-  return pastDests
-}
-
-function sortCurrentDests(currentTrips, dests) {
-  let allDests = dests.destinations
-  let tripIDs = currentTrips.map((trip) => {
-    return trip.destinationID
-  })
-  let currentDests = allDests.filter((dest) => {
-    return tripIDs.includes(dest.id)
-  })
-  return currentDests
-}
 
 function yearlyCost(trips, dests) {
   let targetTrips = trips.filter((trip) => {
@@ -75,16 +49,36 @@ function yearlyCost(trips, dests) {
       totalCost += flightCost + lodgingCost;
     }
   })
-  console.log(totalCost)
   return totalCost
 }
+
+
+//<><><><><>CANNOT TEST<><>><>><<>
+// function sortPastDests(pastTrips, dests) {
+//   let allDests = dests.destinations
+//   let tripIDs = pastTrips.map((trip) => {
+//     return trip.destinationID
+//   })
+//   let pastDests = allDests.filter((dest) => {
+//     return tripIDs.includes(dest.id)
+//   })
+//   yearlyCost(pastTrips, pastDests)
+//   return pastDests
+// }
+
+// function sortCurrentDests(currentTrips, dests) {
+//   let allDests = dests.destinations
+//   let tripIDs = currentTrips.map((trip) => {
+//     return trip.destinationID
+//   })
+//   let currentDests = allDests.filter((dest) => {
+//     return tripIDs.includes(dest.id)
+//   })
+//   return currentDests
+// }
 export {
   userTrips,
-  sortTrips,
   sortPastTrips,
   sortCurrentTrips,
-  sortPastDests,
-  sortCurrentDests,
-  pastTrips,
-  currentTrips,
+  yearlyCost,
 }
